% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallingCNV.R
\name{CallingCNV}
\alias{CallingCNV}
\title{Calling the CNV}
\usage{
CallingCNV(object, modelNames = "E")
}
\arguments{
\item{object}{a CNVcaller object}

\item{modelNames}{A vector of character strings indicating the models to be fitted in the EM phase of clustering.}
}
\value{
object
}
\description{
Calling the CNV
}
\examples{
CallingCNV(object, "E")
}
