% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatObject.R
\name{CNVision}
\alias{CNVision}
\title{Create a CNVision object}
\usage{
CNVision(dir)
}
\arguments{
\item{dir}{Directory containing BAM files}

\item{length}{The length of reads}

\item{genome}{the Reference Genome}

\item{resolution}{the number of bins}
}
\value{
A CNVision S4 object containing cells information and config
}
\description{
Create a CNVision object
}
\examples{
CNVision(150)
}
