% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountRead.R
\name{CountRead}
\alias{CountRead}
\alias{CountRead,CNVisionObj-method}
\title{Count the reads of each bin}
\usage{
CountRead(
  object,
  yieldSize = 1e+05,
  mode = "Union",
  singleEnd = TRUE,
  ignore.strand = TRUE,
  fragments = FALSE,
  ...
)

\S4method{CountRead}{CNVisionObj}(
  object,
  yieldSize = 1e+05,
  mode = "Union",
  singleEnd = TRUE,
  ignore.strand = TRUE,
  fragments = FALSE,
  ...
)
}
\arguments{
\item{object}{a Bin object}

\item{resolution}{the number of bins}
}
\value{
a CNVcaller object
}
\description{
Count the reads of each bin
}
\examples{
CountRead(object)
}
