% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R
\name{detect_peaks}
\alias{detect_peaks}
\title{Title}
\usage{
detect_peaks(
  object,
  peakHeightThreshold = 0.05,
  minpeakdistance = 0.5,
  kernel = "epanechnikov",
  adjust = 1,
  plot = TRUE
)
}
\arguments{
\item{object}{A CNVisionObj}

\item{peakHeightThreshold}{peakHeightThreshold}

\item{kernel}{kernel}

\item{adjust}{adjust}

\item{plot}{plot}

\item{cell}{peaks}
}
\value{
peaks
}
\description{
Title
}
\examples{
peaks
}
