% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_Peaks.R
\name{.detect_peaks}
\alias{.detect_peaks}
\title{Title}
\usage{
.detect_peaks(
  df,
  peakHeightThreshold = 0.05,
  minpeakdistance,
  kernel = "epanechnikov",
  adjust = 1,
  plot = TRUE
)
}
\arguments{
\item{df}{data}

\item{peakHeightThreshold}{peakHeightThreshold}

\item{kernel}{kernel}

\item{adjust}{adjust}

\item{plot}{TRUE}
}
\value{
kde_peaks
}
\description{
Title
}
\examples{
.detect_peaks(x,adjust=1,kernel = "epanechnikov",peakHeightThreshold  = 0.01)
}
