% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{.plotCNV}
\alias{.plotCNV}
\title{Internal CNV Plotting Function}
\usage{
.plotCNV(
  cnv_points,
  cnv_segments,
  xintercept = NULL,
  X_axis_center = NULL,
  annotate_text = NULL,
  annotate_x = NULL,
  annotate_y = NULL,
  ylim_upper = 6,
  without_x = TRUE
)
}
\arguments{
\item{cnv_points}{数据框，包含绘图所需的第一个数据集（例如归一化的 CNV 值和位置）}

\item{cnv_segments}{数据框，包含第二个数据集（例如分段结果）}

\item{xintercept}{数值向量，绘制的垂直间隔线坐标}

\item{X_axis_center}{数值向量，X 轴刻度中点位置}

\item{annotate_text}{字符串，图中注释文本内容}

\item{annotate_x}{数值，注释文本的X坐标}

\item{annotate_y}{数值，注释文本的Y坐标}

\item{ylim_upper}{数值，Y轴上限，用于控制纵轴范围}

\item{data}{数据框，通常是对象中的部分数据，用于辅助绘图}
}
\value{
返回 ggplot2 对象
}
\description{
该函数为内部辅助函数，用于绘制 CNV 相关图形，通常由外部方法调用。
}
\keyword{internal}
