% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R
\name{laplaceMM}
\alias{laplaceMM}
\title{Title}
\usage{
laplaceMM(
  object,
  core_prob = 0.95,
  dist_type = "gaussian",
  background_type = "uniform"
)
}
\arguments{
\item{object}{A CNVisionObj}
}
\value{
model
}
\description{
Title
}
\examples{
model
}
