% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{NormalizeByMeanPlus1}
\alias{NormalizeByMeanPlus1}
\alias{NormalizeData}
\title{Correct Data}
\usage{
NormalizeByMeanPlus1(x)

NormalizeData(object, method = "Normalize")
}
\arguments{
\item{object}{a CNVcller object}

\item{method}{the normalize method}
}
\value{
object
}
\description{
Correct Data
}
\examples{
NormalizeData(object)
}
