% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R, R/Plot.R
\name{plotCNV}
\alias{plotCNV}
\alias{plotCNV,CNVisionObj-method}
\title{plotCNV generic}
\usage{
plotCNV(object, cell, chr = NULL, annotate = TRUE, without_x = TRUE)

\S4method{plotCNV}{CNVisionObj}(object, cell, chr = NULL, annotate = TRUE, without_x = TRUE)
}
\arguments{
\item{object}{CNVisionObj}

\item{...}{other args}
}
\description{
plotCNV generic
}
