% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{Segment}
\alias{Segment}
\title{Identifying the boundaries (physical locations) between genomic regions that have different absolute copy numbers.}
\usage{
Segment(
  object,
  alpha = 0.05,
  nperm = 1000,
  undo.splits = "sdundo",
  undo.SD = 1,
  min.width = 5,
  ...
)
}
\arguments{
\item{object}{a CNVison object}

\item{alpha}{1}

\item{nperm}{1}

\item{undo.SD}{1}

\item{min.width}{1}
}
\value{
object
}
\description{
Identifying the boundaries (physical locations) between genomic regions that have different absolute copy numbers.
}
\examples{
1
}
